+++ src/beans/index.tsx
@@ -34,6 +34,7 @@ export enum LicenseTypeEnum {
   real = "reale",
   virtual = "virtuale",
   trial = "saggio",
+  temp = "reale_temp",
 }
 export type LicenseType = keyof typeof LicenseTypeEnum;
 
@@ -95,16 +96,16 @@ export type TabItem = {
   disabled?: boolean;
 };
 
-export type TabOrientationBean = TabOrientationEnum;
+export type NavigationTabsOrientation = keyof typeof NavigationTabsOrientations;
 
-export enum TabOrientationEnum {
+export enum NavigationTabsOrientations {
   horizontal = "horizontal",
   vertical = "vertical",
 }
 
-export type TabSizeBean = TabSizeEnum;
+export type NavigationTabsSize = keyof typeof NavigationTabsSizes;
 
-export enum TabSizeEnum {
+export enum NavigationTabsSizes {
   big = "big",
   small = "small",
 }
@@ -284,14 +285,6 @@ export enum AvatarSize {
   large = "large",
 }
 
-export enum StatusTagStatus {
-  warning = "warning",
-  error = "error",
-  success = "success",
-  neutral = "neutral",
-  default = "default",
-}
-
 export enum ListSize {
   small = "small",
   medium = "medium",
@@ -434,7 +427,8 @@ export enum ZFileUploadTypeEnum {
 
 export enum ZChipType {
   default = "default",
-  mini = "mini",
+  medium = "medium",
+  small = "small",
 }
 
 export enum ZSectionTitleDividerPositions {
@@ -445,8 +439,33 @@ export enum ZSectionTitleDividerPositions {
 export type ZSectionTitleDividerPosition =
   keyof typeof ZSectionTitleDividerPositions;
 
+export enum ZDatePickerMode {
+  date = "date",
+  dateTime = "date-time",
+  months = "months",
+}
+
+export enum ZDatePickerModeValues {
+  DATE = 401,
+  DATETIME = 523,
+  MONTHS = 288,
+}
+
+export enum ZDatePickerPosition {
+  top = "top",
+  bottom = "bottom",
+}
+
 export declare type SkipToContentLink = {
   label: string;
   href: string;
   ariaLabel?: string;
 };
+
+export enum StatusTagStatus {
+  warning = "warning",
+  error = "error",
+  success = "success",
+  neutral = "neutral",
+  default = "default",
+}