+++ src/components.d.ts
@@ -6,7 +6,7 @@
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
 import { AlertTypes, LicenseTypeEnum, MenuItem as MenuItem1, TooltipPosition } from "./beans/index";
-import { AvatarSize, ButtonSizeEnum, ButtonVariantBean, ButtonVariantEnum, CardVariants, ComboItemBean, DictionaryData, DividerOrientation, DividerSize, ExpandableListButtonAlign, ExpandableListStyle, HeaderUserData, InputStatusBean, InputTypeBean, ListDividerType, ListSize, MenuItem, NotificationType, OffCanvasVariantsEnum, PocketStatus, PopoverBorderRadius, PopoverPosition, PopoverShadow, SelectItemBean, Size, SkipToContentLink, SortDirection, StatusTagStatus, TabOrientationBean, TabSizeBean, ThemeVariant, ThemeVariantBean, ToastNotificationPositionsTypes, ToastNotificationTransitionTypes, ToastNotificationTypes, TransitionDirectionEnum, ZChipType, ZFileUploadTypeEnum, ZSectionTitleDividerPosition, ZTableRowExpandedType, ZtoggleSwitchPositionEnum } from "./beans";
+import { AvatarSize, ButtonSizeEnum, ButtonVariantBean, ButtonVariantEnum, CardVariants, ComboItemBean, DictionaryData, DividerOrientation, DividerSize, ExpandableListButtonAlign, ExpandableListStyle, HeaderUserData, InputStatusBean, InputTypeBean, ListDividerType, ListSize, MenuItem, NavigationTabsOrientation, NavigationTabsSize, NotificationType, OffCanvasVariantsEnum, PocketStatus, PopoverBorderRadius, PopoverPosition, PopoverShadow, SelectItemBean, Size, SkipToContentLink, SortDirection, StatusTagStatus, ThemeVariant, ThemeVariantBean, ToastNotificationPositionsTypes, ToastNotificationTransitionTypes, ToastNotificationTypes, TransitionDirectionEnum, ZChipType, ZDatePickerMode, ZFileUploadTypeEnum, ZSectionTitleDividerPosition, ZTableRowExpandedType, ZtoggleSwitchPositionEnum } from "./beans";
 import { ListItemBean } from "./beans/index.js";
 import { ZTypographyLevels } from "./components/typography/z-typography/index";
 export namespace Components {
@@ -215,18 +215,33 @@ export namespace Components {
         /**
           * sets the height of z-carousel ghost loading, this prop is mandatory when isloading is set to true, as otherwise the component won't show.
          */
-        "ghostloadingheight": string;
+        "ghostLoadingHeight": number;
         /**
           * sets whether the z-carousel is on loading state
          */
-        "isloading": boolean;
+        "isLoading": boolean;
     }
     interface ZChip {
-        "boldtext"?: number;
+        /**
+          * set z-chip as disabled
+         */
         "disabled"?: boolean;
-        "filter"?: boolean;
-        "regulartext"?: string;
+        /**
+          * Non interactive icon
+         */
+        "icon"?: string;
+        /**
+          * z-chip interactive icon
+         */
+        "interactiveIcon"?: string;
+        /**
+          * z-chip size type, can be default, medium or small
+         */
         "type"?: ZChipType;
+    }
+    interface ZChipDeprecated {
+        "boldtext"?: number;
+        "regulartext"?: string;
     }
     interface ZCombobox {
         /**
@@ -317,6 +332,16 @@ export namespace Components {
           * hide cookie bar (optional)
          */
         "hide"?: boolean;
+    }
+    interface ZDatePicker {
+        /**
+          * unique id
+         */
+        "datepickerid": string;
+        /**
+          * [Optional] datepicker mode: date, datetime, only months
+         */
+        "mode": ZDatePickerMode;
     }
     interface ZDivider {
         /**
@@ -908,6 +933,10 @@ export namespace Components {
          */
         "cardtype"?: LicenseTypeEnum;
         /**
+          * display footer custom slotted content
+         */
+        "customContent": boolean;
+        /**
           * faded status
          */
         "faded": boolean;
@@ -1049,20 +1078,82 @@ export namespace Components {
         "userdata"?: string | HeaderUserData;
     }
     interface ZNavigationTab {
+        /**
+          * Whether the tab is disabled.
+         */
         "disabled"?: boolean;
-        "orientation": string;
+        /**
+          * Html `title` attribute for the button.
+         */
+        "htmlTitle": string;
+        /**
+          * Name of the icon to use. Use the slot `icon` for extra customization. The `filled` version will be automatically used (if found) when the tab is `selected`.
+         */
+        "icon": string;
+        /**
+          * Label to show in the tab.
+         */
+        "label": string;
+        /**
+          * Tab orientation. Do not set this manually: `z-navigation-tabs` will handle this.
+         */
+        "orientation": NavigationTabsOrientation;
+        /**
+          * Whether the tab is selected.
+         */
         "selected"?: boolean;
-        "size": string;
+        /**
+          * Tab size. Do not set this manually: `z-navigation-tabs` will handle this.
+         */
+        "size": NavigationTabsSize;
+    }
+    interface ZNavigationTabLink {
+        /**
+          * Whether the tab is disabled.
+         */
+        "disabled"?: boolean;
+        /**
+          * Url to set to the anchor element.
+         */
+        "href": string;
+        /**
+          * Html title attribute for the anchor element.
+         */
+        "htmlTitle": string;
+        /**
+          * Name of the icon to use. Use the slot `icon` for extra customization. The `filled` version will be automatically used (if found) when the tab is `selected`.
+         */
+        "icon": string;
+        /**
+          * Label to show in the tab.
+         */
+        "label": string;
+        /**
+          * Tab orientation. Do not set this manually: `z-navigation-tabs` will handle this.
+         */
+        "orientation": NavigationTabsOrientation;
+        /**
+          * Whether the tab is selected.
+         */
+        "selected"?: boolean;
+        /**
+          * Tab size. Do not set this manually: `z-navigation-tabs` will handle this.
+         */
+        "size": NavigationTabsSize;
+        /**
+          * Html `target` attribute for the anchor element.
+         */
+        "target": string;
     }
     interface ZNavigationTabs {
         /**
-          * Available orientation: `horizontal` and `vertical`. Defaults to `horizontal`.
+          * Navigation tabs orientation.
          */
-        "orientation"?: TabOrientationBean;
+        "orientation"?: NavigationTabsOrientation;
         /**
-          * Available sizes: `big` and `small`. Defaults to `big`.
+          * Navigation tabs size.
          */
-        "size"?: TabSizeBean;
+        "size"?: NavigationTabsSize;
     }
     interface ZNotification {
         /**
@@ -1521,6 +1612,16 @@ export namespace Components {
     }
     interface ZTableStickyFooter {
     }
+    interface ZTag {
+        /**
+          * [optional] Hide the text and show it on hover
+         */
+        "expandable"?: boolean;
+        /**
+          * [optional] Tag icon
+         */
+        "icon"?: string;
+    }
     interface ZToastNotification {
         /**
           * toast notification closing timeout (ms)
@@ -1727,6 +1828,12 @@ declare global {
         prototype: HTMLZChipElement;
         new (): HTMLZChipElement;
     };
+    interface HTMLZChipDeprecatedElement extends Components.ZChipDeprecated, HTMLStencilElement {
+    }
+    var HTMLZChipDeprecatedElement: {
+        prototype: HTMLZChipDeprecatedElement;
+        new (): HTMLZChipDeprecatedElement;
+    };
     interface HTMLZComboboxElement extends Components.ZCombobox, HTMLStencilElement {
     }
     var HTMLZComboboxElement: {
@@ -1744,6 +1851,12 @@ declare global {
     var HTMLZCookiebarElement: {
         prototype: HTMLZCookiebarElement;
         new (): HTMLZCookiebarElement;
+    };
+    interface HTMLZDatePickerElement extends Components.ZDatePicker, HTMLStencilElement {
+    }
+    var HTMLZDatePickerElement: {
+        prototype: HTMLZDatePickerElement;
+        new (): HTMLZDatePickerElement;
     };
     interface HTMLZDividerElement extends Components.ZDivider, HTMLStencilElement {
     }
@@ -1997,6 +2110,12 @@ declare global {
         prototype: HTMLZNavigationTabElement;
         new (): HTMLZNavigationTabElement;
     };
+    interface HTMLZNavigationTabLinkElement extends Components.ZNavigationTabLink, HTMLStencilElement {
+    }
+    var HTMLZNavigationTabLinkElement: {
+        prototype: HTMLZNavigationTabLinkElement;
+        new (): HTMLZNavigationTabLinkElement;
+    };
     interface HTMLZNavigationTabsElement extends Components.ZNavigationTabs, HTMLStencilElement {
     }
     var HTMLZNavigationTabsElement: {
@@ -2177,6 +2296,12 @@ declare global {
         prototype: HTMLZTableStickyFooterElement;
         new (): HTMLZTableStickyFooterElement;
     };
+    interface HTMLZTagElement extends Components.ZTag, HTMLStencilElement {
+    }
+    var HTMLZTagElement: {
+        prototype: HTMLZTagElement;
+        new (): HTMLZTagElement;
+    };
     interface HTMLZToastNotificationElement extends Components.ZToastNotification, HTMLStencilElement {
     }
     var HTMLZToastNotificationElement: {
@@ -2240,9 +2365,11 @@ declare global {
         "z-card": HTMLZCardElement;
         "z-carousel": HTMLZCarouselElement;
         "z-chip": HTMLZChipElement;
+        "z-chip-deprecated": HTMLZChipDeprecatedElement;
         "z-combobox": HTMLZComboboxElement;
         "z-contextual-menu": HTMLZContextualMenuElement;
         "z-cookiebar": HTMLZCookiebarElement;
+        "z-date-picker": HTMLZDatePickerElement;
         "z-divider": HTMLZDividerElement;
         "z-dragdrop-area": HTMLZDragdropAreaElement;
         "z-file": HTMLZFileElement;
@@ -2285,6 +2412,7 @@ declare global {
         "z-myz-list-item": HTMLZMyzListItemElement;
         "z-myz-topbar": HTMLZMyzTopbarElement;
         "z-navigation-tab": HTMLZNavigationTabElement;
+        "z-navigation-tab-link": HTMLZNavigationTabLinkElement;
         "z-navigation-tabs": HTMLZNavigationTabsElement;
         "z-notification": HTMLZNotificationElement;
         "z-offcanvas": HTMLZOffcanvasElement;
@@ -2315,6 +2443,7 @@ declare global {
         "z-table-header-row": HTMLZTableHeaderRowElement;
         "z-table-row": HTMLZTableRowElement;
         "z-table-sticky-footer": HTMLZTableStickyFooterElement;
+        "z-tag": HTMLZTagElement;
         "z-toast-notification": HTMLZToastNotificationElement;
         "z-toast-notification-list": HTMLZToastNotificationListElement;
         "z-toggle-button": HTMLZToggleButtonElement;
@@ -2551,18 +2680,37 @@ declare namespace LocalJSX {
         /**
           * sets the height of z-carousel ghost loading, this prop is mandatory when isloading is set to true, as otherwise the component won't show.
          */
-        "ghostloadingheight"?: string;
+        "ghostLoadingHeight"?: number;
         /**
           * sets whether the z-carousel is on loading state
          */
-        "isloading"?: boolean;
+        "isLoading"?: boolean;
     }
     interface ZChip {
-        "boldtext"?: number;
+        /**
+          * set z-chip as disabled
+         */
         "disabled"?: boolean;
-        "filter"?: boolean;
-        "regulartext"?: string;
+        /**
+          * Non interactive icon
+         */
+        "icon"?: string;
+        /**
+          * z-chip interactive icon
+         */
+        "interactiveIcon"?: string;
+        /**
+          * click on interactive icon
+         */
+        "onInteractiveIconClick"?: (event: CustomEvent<any>) => void;
+        /**
+          * z-chip size type, can be default, medium or small
+         */
         "type"?: ZChipType;
+    }
+    interface ZChipDeprecated {
+        "boldtext"?: number;
+        "regulartext"?: string;
     }
     interface ZCombobox {
         /**
@@ -2665,6 +2813,20 @@ declare namespace LocalJSX {
           * emitted on ACCETTA button click, returns event
          */
         "onAccept"?: (event: CustomEvent<any>) => void;
+    }
+    interface ZDatePicker {
+        /**
+          * unique id
+         */
+        "datepickerid"?: string;
+        /**
+          * [Optional] datepicker mode: date, datetime, only months
+         */
+        "mode"?: ZDatePickerMode;
+        /**
+          * emitted when date changes, returns selected date
+         */
+        "onDateSelect"?: (event: CustomEvent<any>) => void;
     }
     interface ZDivider {
         /**
@@ -3352,6 +3514,10 @@ declare namespace LocalJSX {
          */
         "cardtype"?: LicenseTypeEnum;
         /**
+          * display footer custom slotted content
+         */
+        "customContent"?: boolean;
+        /**
           * faded status
          */
         "faded"?: boolean;
@@ -3505,21 +3671,84 @@ declare namespace LocalJSX {
         "userdata"?: string | HeaderUserData;
     }
     interface ZNavigationTab {
+        /**
+          * Whether the tab is disabled.
+         */
         "disabled"?: boolean;
+        /**
+          * Html `title` attribute for the button.
+         */
+        "htmlTitle"?: string;
+        /**
+          * Name of the icon to use. Use the slot `icon` for extra customization. The `filled` version will be automatically used (if found) when the tab is `selected`.
+         */
+        "icon"?: string;
+        /**
+          * Label to show in the tab.
+         */
+        "label"?: string;
         "onSelected"?: (event: CustomEvent<any>) => void;
-        "orientation"?: string;
+        /**
+          * Tab orientation. Do not set this manually: `z-navigation-tabs` will handle this.
+         */
+        "orientation"?: NavigationTabsOrientation;
+        /**
+          * Whether the tab is selected.
+         */
         "selected"?: boolean;
-        "size"?: string;
+        /**
+          * Tab size. Do not set this manually: `z-navigation-tabs` will handle this.
+         */
+        "size"?: NavigationTabsSize;
+    }
+    interface ZNavigationTabLink {
+        /**
+          * Whether the tab is disabled.
+         */
+        "disabled"?: boolean;
+        /**
+          * Url to set to the anchor element.
+         */
+        "href"?: string;
+        /**
+          * Html title attribute for the anchor element.
+         */
+        "htmlTitle"?: string;
+        /**
+          * Name of the icon to use. Use the slot `icon` for extra customization. The `filled` version will be automatically used (if found) when the tab is `selected`.
+         */
+        "icon"?: string;
+        /**
+          * Label to show in the tab.
+         */
+        "label"?: string;
+        "onSelected"?: (event: CustomEvent<any>) => void;
+        /**
+          * Tab orientation. Do not set this manually: `z-navigation-tabs` will handle this.
+         */
+        "orientation"?: NavigationTabsOrientation;
+        /**
+          * Whether the tab is selected.
+         */
+        "selected"?: boolean;
+        /**
+          * Tab size. Do not set this manually: `z-navigation-tabs` will handle this.
+         */
+        "size"?: NavigationTabsSize;
+        /**
+          * Html `target` attribute for the anchor element.
+         */
+        "target"?: string;
     }
     interface ZNavigationTabs {
         /**
-          * Available orientation: `horizontal` and `vertical`. Defaults to `horizontal`.
+          * Navigation tabs orientation.
          */
-        "orientation"?: TabOrientationBean;
+        "orientation"?: NavigationTabsOrientation;
         /**
-          * Available sizes: `big` and `small`. Defaults to `big`.
+          * Navigation tabs size.
          */
-        "size"?: TabSizeBean;
+        "size"?: NavigationTabsSize;
     }
     interface ZNotification {
         /**
@@ -3735,6 +3964,10 @@ declare namespace LocalJSX {
           * [optional] Box shadow token for popover
          */
         "boxShadow"?: PopoverShadow;
+        /**
+          * Emitted on popover click, returns isVisible state
+         */
+        "onTriggerClick"?: (event: CustomEvent<any>) => void;
         /**
           * [optional] Sets padding for Popover container
          */
@@ -4012,6 +4245,16 @@ declare namespace LocalJSX {
     }
     interface ZTableStickyFooter {
     }
+    interface ZTag {
+        /**
+          * [optional] Hide the text and show it on hover
+         */
+        "expandable"?: boolean;
+        /**
+          * [optional] Tag icon
+         */
+        "icon"?: string;
+    }
     interface ZToastNotification {
         /**
           * toast notification closing timeout (ms)
@@ -4168,9 +4411,11 @@ declare namespace LocalJSX {
         "z-card": ZCard;
         "z-carousel": ZCarousel;
         "z-chip": ZChip;
+        "z-chip-deprecated": ZChipDeprecated;
         "z-combobox": ZCombobox;
         "z-contextual-menu": ZContextualMenu;
         "z-cookiebar": ZCookiebar;
+        "z-date-picker": ZDatePicker;
         "z-divider": ZDivider;
         "z-dragdrop-area": ZDragdropArea;
         "z-file": ZFile;
@@ -4213,6 +4458,7 @@ declare namespace LocalJSX {
         "z-myz-list-item": ZMyzListItem;
         "z-myz-topbar": ZMyzTopbar;
         "z-navigation-tab": ZNavigationTab;
+        "z-navigation-tab-link": ZNavigationTabLink;
         "z-navigation-tabs": ZNavigationTabs;
         "z-notification": ZNotification;
         "z-offcanvas": ZOffcanvas;
@@ -4243,6 +4489,7 @@ declare namespace LocalJSX {
         "z-table-header-row": ZTableHeaderRow;
         "z-table-row": ZTableRow;
         "z-table-sticky-footer": ZTableStickyFooter;
+        "z-tag": ZTag;
         "z-toast-notification": ZToastNotification;
         "z-toast-notification-list": ZToastNotificationList;
         "z-toggle-button": ZToggleButton;
@@ -4271,9 +4518,11 @@ declare module "@stencil/core" {
             "z-card": LocalJSX.ZCard & JSXBase.HTMLAttributes<HTMLZCardElement>;
             "z-carousel": LocalJSX.ZCarousel & JSXBase.HTMLAttributes<HTMLZCarouselElement>;
             "z-chip": LocalJSX.ZChip & JSXBase.HTMLAttributes<HTMLZChipElement>;
+            "z-chip-deprecated": LocalJSX.ZChipDeprecated & JSXBase.HTMLAttributes<HTMLZChipDeprecatedElement>;
             "z-combobox": LocalJSX.ZCombobox & JSXBase.HTMLAttributes<HTMLZComboboxElement>;
             "z-contextual-menu": LocalJSX.ZContextualMenu & JSXBase.HTMLAttributes<HTMLZContextualMenuElement>;
             "z-cookiebar": LocalJSX.ZCookiebar & JSXBase.HTMLAttributes<HTMLZCookiebarElement>;
+            "z-date-picker": LocalJSX.ZDatePicker & JSXBase.HTMLAttributes<HTMLZDatePickerElement>;
             "z-divider": LocalJSX.ZDivider & JSXBase.HTMLAttributes<HTMLZDividerElement>;
             "z-dragdrop-area": LocalJSX.ZDragdropArea & JSXBase.HTMLAttributes<HTMLZDragdropAreaElement>;
             "z-file": LocalJSX.ZFile & JSXBase.HTMLAttributes<HTMLZFileElement>;
@@ -4316,6 +4565,7 @@ declare module "@stencil/core" {
             "z-myz-list-item": LocalJSX.ZMyzListItem & JSXBase.HTMLAttributes<HTMLZMyzListItemElement>;
             "z-myz-topbar": LocalJSX.ZMyzTopbar & JSXBase.HTMLAttributes<HTMLZMyzTopbarElement>;
             "z-navigation-tab": LocalJSX.ZNavigationTab & JSXBase.HTMLAttributes<HTMLZNavigationTabElement>;
+            "z-navigation-tab-link": LocalJSX.ZNavigationTabLink & JSXBase.HTMLAttributes<HTMLZNavigationTabLinkElement>;
             "z-navigation-tabs": LocalJSX.ZNavigationTabs & JSXBase.HTMLAttributes<HTMLZNavigationTabsElement>;
             "z-notification": LocalJSX.ZNotification & JSXBase.HTMLAttributes<HTMLZNotificationElement>;
             "z-offcanvas": LocalJSX.ZOffcanvas & JSXBase.HTMLAttributes<HTMLZOffcanvasElement>;
@@ -4346,6 +4596,7 @@ declare module "@stencil/core" {
             "z-table-header-row": LocalJSX.ZTableHeaderRow & JSXBase.HTMLAttributes<HTMLZTableHeaderRowElement>;
             "z-table-row": LocalJSX.ZTableRow & JSXBase.HTMLAttributes<HTMLZTableRowElement>;
             "z-table-sticky-footer": LocalJSX.ZTableStickyFooter & JSXBase.HTMLAttributes<HTMLZTableStickyFooterElement>;
+            "z-tag": LocalJSX.ZTag & JSXBase.HTMLAttributes<HTMLZTagElement>;
             "z-toast-notification": LocalJSX.ZToastNotification & JSXBase.HTMLAttributes<HTMLZToastNotificationElement>;
             "z-toast-notification-list": LocalJSX.ZToastNotificationList & JSXBase.HTMLAttributes<HTMLZToastNotificationListElement>;
             "z-toggle-button": LocalJSX.ZToggleButton & JSXBase.HTMLAttributes<HTMLZToggleButtonElement>;